back/requirement.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.0
astroid==2.4.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==4.1.1
certifi==2020.11.8
cfgv==3.2.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.0rc1
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
dj-rest-auth==1.1.2
Django==3.1.3
django-allauth==0.43.0
django-braces==1.14.0
django-filter==2.4.0
django-heroku==0.3.1
django-nose==1.4.7
django-oauth-toolkit==1.3.3
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.5.0
dodgy==0.2.1
drf-yasg==1.20.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
flake8==3.8.4
flake8-polyfill==1.0.2
future==0.18.2
GDAL==3.0.4
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
identify==1.5.10
idna==2.10
importlib-metadata==1.5.0
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
nodeenv==1.5.0
nose==1.3.7
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandas==1.2.0
paramiko==2.6.0
pep8-naming==0.10.0
pexpect==4.6.0
Pillow==7.0.0
pinocchio==0.4.2
pre-commit==2.9.2
prospector==1.3.1
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycups==1.9.73
pydocstyle==5.1.1
pyflakes==2.2.0
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.1.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pypi-install==0.0.5
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2020.4
pyxdg==0.26
PyYAML==5.3.1
rednose==1.3.0
reportlab==3.5.34
requests==2.25.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
requirements-detector==0.7
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
SecretStorage==2.3.1
setoptconf==0.2.0
simplejson==3.16.0
six==1.15.0
snowballstemmer==2.0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
systemd-python==234
termstyle==0.1.11
toml==0.10.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.2
usb-creator==0.3.7
virtualenv==20.0.35
wadllib==1.3.3
whitenoise==5.2.0
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
django-cors-headers==3.7.0